Refactor color and font slug mapping to TS
[MAILPOET-3005]
